+++ NEWS.md
@@ -1,5 +1,6 @@
 ### Unreleased 1.8 patch
 
+* Improved rendering quality of the Mini-map when it's small (#1431)
 * Fixed automatic tool switching after deleting layers
 * Fixed map bounding rectangle for infinite isometric maps
 * Fixed tile selection to not get removed when deleting (#3281)